requirements.txt
streamlit
langchain
langchain-groq
langchain-community
python-dotenv
pypdf
faiss-cpu
sqlalchemy
pandas
email-validator
